security improvements in methods and publications

+++ imports/api/clients/methods.js
@@ -1,4 +1,5 @@
 import { logger } from '/imports/utils/logger';
+import { ClientsAccessRoles } from '/imports/config/permissions';
 import { PendingClients } from '/imports/api/pendingClients/pendingClients';
 import { HmisClient } from '/imports/api/hmisApi';
 import { mergeByDedupId } from '/imports/api/clients/helpers';
@@ -9,8 +10,13 @@ import eventPublisher, {
 
 Meteor.methods({
   'clients.create'(client, schema = 'v2017', clientVersion = false) {
-    logger.info(`METHOD[${Meteor.userId()}]: clients.create`, client);
-    const hc = HmisClient.create(Meteor.userId());
+    logger.info(`METHOD[${this.userId}]: clients.create`, client);
+
+    if (!Roles.userIsInRole(this.userId, ClientsAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
+
+    const hc = HmisClient.create(this.userId);
     const result = hc.api('client').createClient(client, schema);
 
     if (clientVersion) return result;
@@ -28,10 +34,15 @@ Meteor.methods({
   },
 
   'clients.update'(clientId, client, schema) {
-    logger.info(`METHOD[${Meteor.userId()}]: clients.update`, clientId, schema, client);
+    logger.info(`METHOD[${this.userId}]: clients.update`, clientId, schema, client);
+
     check(clientId, String);
     check(schema, String);
-    const hc = HmisClient.create(Meteor.userId());
+    if (!Roles.userIsInRole(this.userId, ClientsAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
+
+    const hc = HmisClient.create(this.userId);
     hc.api('client').updateClient(clientId, client, schema);
 
     eventPublisher.publish(new ClientUpdatedEvent({
@@ -45,46 +56,48 @@ Meteor.methods({
 
   'clients.delete'(clientId, schema) { // eslint-disable-line
     logger.info(`METHOD[${Meteor.userId()}]: clients.delete`, clientId);
+
     check(clientId, String);
     check(schema, String);
+    if (!Roles.userIsInRole(this.userId, ClientsAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
+
     const hc = HmisClient.create(Meteor.userId());
     hc.api('client').deleteClient(clientId, schema);
   },
 
-  updateClientMatchStatus(clientId, statusCode, comments, recipients) {
-    const hc = HmisClient.create(Meteor.userId());
+  'clients.updateMatchStatus'(clientId, statusCode, comments, recipients) {
+    logger.info(`METHOD[${this.userId}]: clients.updateMatchStatus`,
+      clientId, statusCode, comments, recipients
+    );
+
+    check(clientId, String);
+    if (!Roles.userIsInRole(this.userId, ClientsAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
+
+    const hc = HmisClient.create(this.userId);
     return hc.api('house-matching').updateClientMatchStatus(
       clientId, statusCode, comments, recipients
     );
   },
 
-  saveToSchema(client, inputSchema) {
-    // Creates a client version if it doesn't exist for this version
-    const { clientVersions } = client;
-    // Check if exists:
-    const clientVersion = clientVersions.find(({ schema }) => schema === inputSchema);
-    if (clientVersion) return clientVersion;
+  async searchClient(query, options = {}) {
+    logger.info(`METHOD[${this.userId}]: searchClient(${query})`);
+    if (!Roles.userIsInRole(this.userId, ClientsAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
 
-    const hc = HmisClient.create(Meteor.userId());
-    return hc.api('client').createClient({
-      ...client, suffix: client.nameSuffix || '',
-    }, inputSchema); // { clientId, schema }
-  },
+    const limit = Math.min(options.limit, 50);
 
-  async searchClient(query, options) {
-    logger.info(`METHOD[${Meteor.userId()}]: searchClient(${query})`);
-    const optionz = options || {};
-
-    // guard against client-side DOS: hard limit to 50
-    optionz.limit = Math.min(50, Math.abs(optionz.limit || 50));
-
-    const hc = HmisClient.create(Meteor.userId());
-    let hmisClients = hc.api('client').searchClient(query, optionz.limit);
+    const hc = HmisClient.create(this.userId);
+    let hmisClients = hc.api('client').searchClient(query, limit);
 
     hmisClients = hmisClients.filter(client => client.link);
 
     let localClients = [];
-    if (!optionz.excludeLocalClients) {
+    if (!options.excludeLocalClients) {
       try {
         localClients = PendingClients.aggregate([
           {
@@ -116,7 +129,7 @@ Meteor.methods({
             },
           },
           {
-            $limit: optionz.limit,
+            $limit: limit,
           },
         ], { explain: false });
       } catch (err) {
@@ -131,7 +144,6 @@ Meteor.methods({
           if (localClients[i].personalId === hmisClients[j].clientId) {
             // Remove.
             localClients.splice(i, 1);
-            logger.info('Element Removed');
             break;
           }
         }
