+++ imports/api/files/methods.js
@@ -1,17 +1,23 @@
 import Files from '/imports/api/files/files';
 import { logger } from '/imports/utils/logger';
+import { FilesAccessRoles } from '/imports/config/permissions';
 
 Meteor.methods({
   'files.create'(doc) {
     logger.info(`METHOD[${Meteor.userId()}]: files.create`, doc);
     check(doc, Files.schema);
-    // TODO: permissions
+    if (!Roles.userIsInRole(this.userId, FilesAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
     return Files.insert(doc);
   },
+
   'files.delete'(id) {
     logger.info(`METHOD[${Meteor.userId()}]: files.delete`, id);
     check(id, String);
-    // TODO: permissions
+    if (!Roles.userIsInRole(this.userId, FilesAccessRoles)) {
+      throw new Meteor.Error(403, 'Forbidden');
+    }
     const currentFile = Files.findOne(id);
     Files.Uploads.remove(currentFile.fileId);
     Files.remove(id);
